+++ src/view.cpp
@@ -168,7 +168,7 @@ std::u32string View::IntToSmuflFigures(unsigned short number, int offset)
     stream << number;
     std::u32string str = UTF8to32(stream.str());
 
-    for (size_t i = 0; i < str.size(); ++i) {
+    for (int i = 0; i < (int)str.size(); ++i) {
         str[i] += offset - 48;
     }
     return str;
